Relocations ftw
Doesn't work yet.
