+++ src/monodroid/jni/xamarin-app.hh
@@ -10,7 +10,7 @@
 #include "monodroid.h"
 #include "xxhash.hh"
 
-static constexpr uint64_t FORMAT_TAG = 0x015E6972616D58;
+static constexpr uint64_t FORMAT_TAG = 0x00025E6972616D58; // 'Xmari^XY' where XY is the format version
 static constexpr uint32_t COMPRESSED_DATA_MAGIC = 0x5A4C4158; // 'XALZ', little-endian
 static constexpr uint32_t ASSEMBLY_STORE_MAGIC = 0x41424158; // 'XABA', little-endian
 static constexpr uint32_t ASSEMBLY_STORE_FORMAT_VERSION = 1; // Increase whenever an incompatible change is made to the
